+++ src/main/java/kr/codesqaud/cafe/controller/UserController.java
@@ -3,9 +3,9 @@ package kr.codesqaud.cafe.controller;
 import kr.codesqaud.cafe.dto.user.UserLoginDTO;
 import kr.codesqaud.cafe.repository.member.MemberRepository;
 import kr.codesqaud.cafe.domain.User;
-import kr.codesqaud.cafe.validation.UserJoinValidator;
-import kr.codesqaud.cafe.validation.UserLoginValidator;
-import kr.codesqaud.cafe.validation.UserUpdateValidator;
+import kr.codesqaud.cafe.validation.user.UserJoinValidator;
+import kr.codesqaud.cafe.validation.user.UserLoginValidator;
+import kr.codesqaud.cafe.validation.user.UserUpdateValidator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
